+++ tests/testthat/test-ImposeConstraint.R
@@ -27,8 +27,9 @@ test_that("ImposeConstraint() works", {
     c(0, 0, '?', '?', 1, 1,
       1, 1,   1, '?', 0, 0), ncol = 2,
     dimnames = list(letters[1:6], NULL)))
-  expect_true(inherits(expect_warning(
-    ImposeConstraint(NJTree(constraint), constraint)), 'phylo'))
+
+  expect_warning(out <- ImposeConstraint(NJTree(constraint), constraint))
+  expect_true(inherits(out, 'phylo'))
 
   # Need to collapse splits intelligently to avoid error.  From Joe Moysiuk.
   constraint <- MatrixToPhyDat(